Add files via upload
This code will first read the data of the video game sales .csv file into a data frame using pandas. Then, the data will be sorted by (country_Sales) and (Genre), descending. This will show the top genres sold in each country. This is done for Japan, North America, and Europe. The matplotlib.pyplot (shortened as plt) function is used to create a scatter plot for each country.