Add files via upload
This code will first read the data of the video game sales .csv file into a handle named 'data.' A for loop created for that sorted data will append the first 5 most popular games in global sales. This will show the top games sold in global sales. The matplotlib.pyplot (shortened as plt) function is used to create a histogram that shows the global top 5 games.