Add files via upload
This code will first read the data of the video game sales .csv file into a dataframe named 'df.' A variable named 'sum' is made of a .groupby dataframe, consisting of (Platform) and (Global_Sales). The sum of (Global_Sales) is taken. The matplotlib.pyplot (shortened as plt) function is used to create a pie chart that shows the percentages that each platform has in terms of (Global_Sales).