Add files via upload
This code will first read the data of the video game sales .csv file into a .reader named data. Then, a counter is established to stop the for loop from appending more than 10 games to the x and y lists. The pyplot function from matplotlib (plt) will be used to create a lollipop plot. This will show the top 10 games sold in North America.